fix: LHS of AND gates
We now store the LHS of AND gates on-demand when/if we discover they are not
computable. In parsing a binary AIG file, the LHS of all AND gates is assumed to
be inferable, though we haven't actually implemented this parsing yet. In an
ASCII AIG file, the LHS of AND gates is assumed to be inferable until we
discover one that is not. At that point we construct storage for the LHS of AND
gates and treat this as authoritative from then on.

Github: closes #1 "LHS of AND gates is wrong"